requirements.txt
supabase
pytest
requests
python-dotenv
faker